+++ CHANGELOG.md
@@ -5,6 +5,9 @@
 - Each breaking change should be preceded with `**BREAKING CHANGE**:` in the changelog entry. The minor version (major when we hit v1.0.0) should also increase accordingly. If it's an upcoming release, put the next version in brackets.
 - When the new version is ready, rename the header to the version number. After that, create the according tag. Also make sure to change the package.json version as well.
 
+## Upcoming Release
+- Renamed Generator group in `TMAGTerrainGrid` to Graph
+
 ## v0.5.0
 - Texture Node now uses the smallest of dimensions instead of width for calculating input resolution. This isn't considered breaking as this node is meant to be used with heightmaps or splatmaps, i.e. textures that are usually square
 - Added check that compute shaders and random write textures are actually supported